+++ eqcorrscan/core/match_filter/party.py
@@ -34,6 +34,7 @@ from eqcorrscan.core.match_filter.matched_filter import (
     MatchFilterError, _group_process)
 from eqcorrscan.core.match_filter.template import Template
 from eqcorrscan.core.match_filter.family import Family
+from eqcorrscan.core.match_filter.detection import write_detections
 from eqcorrscan.core.match_filter.helpers import (
     _total_microsec, temporary_directory, _safemembers, _templates_match)
 
@@ -635,8 +636,8 @@ class Party(object):
                 raise MatchFilterError(
                     'Will not overwrite existing file: %s' % filename)
             for family in self.families:
-                for detection in family.detections:
-                    detection.write(fname=filename, append=True)
+                write_detections(fname=filename, detections=family.detections,
+                                 mode="a")
         elif format.lower() == 'tar':
             if os.path.exists(filename):
                 raise IOError('Will not overwrite existing file: %s'
@@ -742,7 +743,8 @@ class Party(object):
     def lag_calc(self, stream, pre_processed, shift_len=0.2, min_cc=0.4,
                  horizontal_chans=['E', 'N', '1', '2'], vertical_chans=['Z'],
                  cores=1, interpolate=False, plot=False, parallel=True,
-                 overlap='calculate', process_cores=None):
+                 overlap='calculate', process_cores=None,
+                 ignore_bad_data=False):
         """
         Compute picks based on cross-correlation alignment.
 
@@ -792,6 +794,11 @@ class Party(object):
         :param process_cores:
             Number of processes to use for pre-processing (if different to
             `cores`).
+        :type ignore_bad_data: bool
+        :param ignore_bad_data:
+            If False (default), errors will be raised if data are excessively
+            gappy or are mostly zeros. If True then no error will be raised,
+            but an empty trace will be returned (and not used in detection).
 
         :returns:
             Catalog of events with picks.  No origin information is included.
@@ -872,7 +879,8 @@ class Party(object):
                 processed_streams = _group_process(
                     template_group=group, cores=process_cores,
                     parallel=parallel, stream=stream.copy(), daylong=False,
-                    ignore_length=False, overlap=lap)
+                    ignore_length=False, overlap=lap,
+                    ignore_bad_data=ignore_bad_data)
                 processed_stream = Stream()
                 for p in processed_streams:
                     processed_stream += p